0428-1 1、用户服务更新 1.1 细化权限服务内容，完善权限树生成，对接网关进行权限验证 1.2 修复Redis存入过程中序列化不一致导致网关无法读取内容的BUG 2、网关更新 2.1 优化URI访问拦截逻辑 2.2 完善用户鉴权模块

+++ UserService/src/main/java/com/q7w/Service/impl/RoleResourceServiceimpl.java
@@ -2,12 +2,22 @@ package com.q7w.Service.impl;
 
 import com.q7w.DAO.RoleResourceDao;
 import com.q7w.Entity.Resource;
+import com.q7w.Entity.Role;
 import com.q7w.Entity.RoleResource;
+import com.q7w.Service.ResourceService;
 import com.q7w.Service.RoleResourceService;
+import com.q7w.Service.RoleService;
+import com.q7w.common.constant.AuthConstant;
+import com.q7w.common.service.RedisService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.TreeMap;
+import java.util.stream.Collectors;
 
 /**
  * @author xiaogu
@@ -17,19 +27,59 @@ import java.util.List;
 public class RoleResourceServiceimpl implements RoleResourceService {
     @Autowired
     RoleResourceDao roleResourceDao;
+    @Autowired
+    RedisService redisService;
+    @Autowired
+    RoleService roleService;
+    @Autowired
+    ResourceService resourceService;
+    @Value("${spring.application.name}")
+    private String applicationName;
     @Override
     public List<RoleResource> findAllByRid(Long rid) {
         return roleResourceDao.findAllByRid(rid);
     }
 
     @Override
-    public void savePermChanges(Long rid, List<Resource> resources) {
-        roleResourceDao.deleteAllByRid(rid);
-        resources.forEach(p -> {
+    public void savePermChanges(Long rid, List<Long> resourceIds) {
+        if (roleResourceDao.findAllByRid(rid).size()!=0){
+            roleResourceDao.deleteAllByRid(rid);
+        }
+        resourceIds.forEach(p -> {
             RoleResource rr = new RoleResource();
             rr.setRid(rid);
-            rr.setPid(p.getId());
+            rr.setPid(p);
             roleResourceDao.save(rr);
         });
+        initResourceRolesMap();
+    }
+
+    public Map<String,List<String>> initResourceRolesMap() {
+        Map<String,List<String>> resourceRoleMap = new TreeMap<>();
+        List<Resource> resourceList = resourceService.listAll();
+        List<Role> roleList = roleService.list();
+        List<RoleResource> relationList = roleResourceDao.findAll();
+        for (Resource resource : resourceList) {
+            Set<Long> roleIds = relationList.stream().filter(item -> item.getPid().equals(resource.getId())).map(RoleResource::getRid).collect(Collectors.toSet());
+            List<String> roleNames = roleList.stream().filter(item -> roleIds.contains(item.getId())).map(item -> item.getId() + "_" + item.getName()).collect(Collectors.toList());
+            resourceRoleMap.put("/"+applicationName+resource.getUrl(),roleNames);
+        }
+        redisService.del(AuthConstant.RESOURCE_ROLES_MAP_KEY);
+        redisService.hSetAll(AuthConstant.RESOURCE_ROLES_MAP_KEY, resourceRoleMap);
+        return resourceRoleMap;
+    }
+
+    @Override
+    public int delresource(Long resid) {
+        roleResourceDao.deleteAllByPid(resid);
+        initResourceRolesMap();
+        return 1;
+    }
+
+    @Override
+    public int delrole(Long rid) {
+        roleResourceDao.deleteAllByRid(rid);
+        initResourceRolesMap();
+        return 1;
     }
 }
